requirements.txt
uvicorn[standard]
pydantic    
python-dotenv
fastmcp
azure-ai-projects
msgraph-sdk
azure-identity